+++ CHANGELOG.md
@@ -7,8 +7,11 @@
 1. [](#improved)
     * If login on registration or activation has been turned on, use login redirect if override is not set
     * Don’t set default templates for `register` and `unauthorized`, use overridable templates [#179](https://github.com/getgrav/grav-plugin-login/issues/179)
+    * Updated `de.yaml` [#175](https://github.com/getgrav/grav-plugin-login/pull/175)
+    * Updated `ru.yaml` [#176](https://github.com/getgrav/grav-plugin-login/pull/176)
 1. [](#bugfix)
     * Fixed broken remember me functionality
+    * Fix uppercase and Unicode username handling [#177](https://github.com/getgrav/grav-plugin-login/pull/177)
 
 # v2.7.3
 ## 06/20/2018